merge dev to main

+++ public/assets/js/objects.js
@@ -22,20 +22,33 @@ document.addEventListener('DOMContentLoaded', function () {
             // inventory.innerHTML = keyImage;
 
             // <input type="submit" value="🔑"></input>
+            const keyExistsInInventory = inventory.querySelector('[name="inventory_item"][value="key"]');
 
-            const inventoryForm = `
-                <form method="post" class="inventory">
-                    <input type="hidden" name="inventory_item" value="key">
-                    <input type="image" src="/assets/images/cle.png" alt="clé" width="50px"></input>
-                </form>
-            `;
-            inventory.innerHTML = inventoryForm;
+            if (!keyExistsInInventory) {
+                const inventoryForm = `
+                    <form method="post" class="inventory">
+                        <input type="hidden" name="inventory_item" value="key">
+                        <input type="image" src="/assets/images/cle.png" alt="clé" width="50px"></input>
+                    </form>
+                `;
+                inventory.innerHTML = inventoryForm;
 
-            // Stocker les données d'inventaire dans la session
-            sessionStorage.setItem('gameInventory', inventoryForm);
+                // Stocker les données d'inventaire dans la session
+                sessionStorage.setItem('gameInventory', inventoryForm);
 
-            // Requête fetch pour envoyer les données mises à jour au serveur
-            updateInventoryOnServer(inventoryForm);
+                // Requête fetch pour envoyer les données mises à jour au serveur
+                updateInventoryOnServer(inventoryForm);
+            }
+        });
+    }
+    const logoutButton = document.querySelector('#logoutButton');
+
+    if (logoutButton) {
+        logoutButton.addEventListener('click', function () {
+            // Remove the inventory data from sessionStorage
+            sessionStorage.removeItem('gameInventory');
+            // Ajoutez d'autres étapes de déconnexion si nécessaire
         });
     }
 });
+
